+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- Add specialized `ILRegisterAccessError` exception for improved error handling on aborted register read/write operations
 - Create config files from DictionaryV3 defaults
 - Override configuration file values with values from another configuration file
 - Allow to upload configurations from configuration file class